requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.3
pytest-django==4.9.0
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.3
responses==0.25.3
sniffio==1.3.1
sqlparse==0.5.2
urllib3==2.2.3
django-environ>=0.7,<1.0
redis
